beamer: tests: move command-related utilities to toplevel util module
These will be needed by other tests too.
